requirements.txt
ExifRead==2.1.2
python-dateutil==2.5.3
six==1.10.0
